eai-portal-support-vaadin/src/test/java/de/unioninvestment/eai/portal/support/vaadin/support/TranslatedDateFieldTest.java
/*
 * Licensed to the Apache Software Foundation (ASF) under one
 * or more contributor license agreements.  See the NOTICE file
 * distributed with this work for additional information
 * regarding copyright ownership.  The ASF licenses this file
 * to you under the Apache License, Version 2.0 (the
 * "License"); you may not use this file except in compliance
 * with the License.  You may obtain a copy of the License at
 *
 *   http://www.apache.org/licenses/LICENSE-2.0
 *
 * Unless required by applicable law or agreed to in writing,
 * software distributed under the License is distributed on an
 * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
 * KIND, either express or implied.  See the License for the
 * specific language governing permissions and limitations
 * under the License.
 */
package de.unioninvestment.eai.portal.support.vaadin.support;

import static org.hamcrest.CoreMatchers.is;
import static org.junit.Assert.assertThat;
import static org.mockito.Mockito.verify;
import static org.mockito.Mockito.when;

import java.util.Date;

import org.junit.Before;
import org.junit.Test;
import org.mockito.Mock;
import org.mockito.MockitoAnnotations;
import org.mockito.invocation.InvocationOnMock;
import org.mockito.stubbing.Answer;
import org.vaadin.addon.propertytranslator.PropertyTranslator;

import com.vaadin.data.Property;

public class TranslatedDateFieldTest {

	@Mock
	private PropertyTranslator translatorMock;
	@Mock
	private Property backendDataSourceMock;

	@Before
	public void setUp() {
		MockitoAnnotations.initMocks(this);
	}

	@Test
	public void shouldInjectTranslatorAroundProperty() {
		when(translatorMock.getType()).thenAnswer(new Answer<Object>() {
			@Override
			public Object answer(InvocationOnMock invocation) throws Throwable {
				return Date.class;
			}
		});
		TranslatedDateField field = new TranslatedDateField(translatorMock);

		field.setPropertyDataSource(backendDataSourceMock);

		verify(translatorMock).setPropertyDataSource(backendDataSourceMock);
		assertThat(field.getPropertyDataSource(), is((Property) translatorMock));
	}
}
